requirements.txt
Django==1.7
MySQL-python==1.2.5
Pillow==2.7.0
argparse==1.2.1
django-autoslug==1.7.2
djangorestframework==3.1.0
pytils==0.3
sorl-thumbnail==12.2
wsgiref==0.1.2
